requirements.txt
pact-python == 2.3.1
pytest == 8.3.5
flask-sqlalchemy==3.1.1
flasgger == 0.9.7.1
PyYAML == 6.0.2
APIFlask == 2.4.0
prance == 25.4.8.0
openapi-spec-validator == 0.7.1
